upgrade context
