+++ package/ddoc-editor.tsx
@@ -24,42 +24,10 @@ import * as Y from 'yjs';
 import MobileToolbar from './components/mobile-toolbar';
 import { fromUint8Array, toUint8Array } from 'js-base64';
 import { PresentationMode } from './components/presentation-mode/presentation-mode';
-import uuid from 'react-uuid';
-import { IComment } from './extensions/comment';
 import { CommentDrawer } from './components/inline-comment/comment-drawer';
 import { CommentBubbleMenu } from './components/inline-comment/comment-bubble-menu';
 import { useResponsive } from './utils/responsive';
-
-const handleAddReply = (
-  comments: IComment[],
-  activeCommentId: string,
-  replyContent: string,
-  setComments: (comments: IComment[]) => void,
-) => {
-  if (!replyContent.trim()) return;
-
-  setComments(
-    comments.map((comment) => {
-      if (comment.id === activeCommentId) {
-        return {
-          ...comment,
-          replies: [
-            ...comment.replies,
-            {
-              id: `reply-${uuid()}`,
-              content: replyContent,
-              replies: [],
-              createdAt: new Date(),
-              selectedContent: comment.selectedContent,
-            },
-          ],
-          content: comment.content,
-        };
-      }
-      return comment;
-    }),
-  );
-};
+import { CommentProvider } from './components/inline-comment/context/comment-context';
 
 const DdocEditor = forwardRef(
   (
@@ -142,14 +110,10 @@ const DdocEditor = forwardRef(
       ref: editorRef,
       isContentLoading,
       ydoc,
-      comments,
+      refreshYjsIndexedDbProvider,
       activeCommentId,
       setActiveCommentId,
-      setComments,
-      commentsSectionRef,
-      setComment,
-      focusCommentInEditor,
-      refreshYjsIndexedDbProvider,
+      focusCommentWithActiveId,
     } = useDdocEditor({
       enableIndexeddbSync,
       ddocId,
@@ -345,269 +309,244 @@ const DdocEditor = forwardRef(
         >
           {renderNavbar?.({ editor: editor.getJSON() })}
         </nav>
-        {!isPreviewMode && (
+        <CommentProvider
+          editor={editor}
+          username={username as string}
+          walletAddress={walletAddress as string}
+          activeCommentId={activeCommentId}
+          setActiveCommentId={setActiveCommentId}
+          focusCommentWithActiveId={focusCommentWithActiveId}
+          initialComments={[]}
+        >
+          {!isPreviewMode && (
+            <div
+              id="toolbar"
+              className={cn(
+                'z-50 hidden xl:flex items-center justify-center w-full h-[52px] fixed left-0 px-1 bg-[#ffffff] border-b color-border-default transition-transform duration-300 top-[3.5rem]',
+                {
+                  'translate-y-0': isNavbarVisible,
+                  'translate-y-[-108%]': !isNavbarVisible,
+                },
+              )}
+            >
+              <div className="justify-center items-center grow relative">
+                <EditorToolBar
+                  onError={onError}
+                  editor={editor}
+                  zoomLevel={zoomLevel}
+                  setZoomLevel={setZoomLevel}
+                  isNavbarVisible={isNavbarVisible}
+                  setIsNavbarVisible={setIsNavbarVisible}
+                  secureImageUploadUrl={secureImageUploadUrl}
+                  onMarkdownExport={onMarkdownExport}
+                  onMarkdownImport={onMarkdownImport}
+                />
+              </div>
+            </div>
+          )}
+          {isPresentationMode && (
+            <PresentationMode
+              editor={editor}
+              onClose={handleClosePresentationMode}
+              isFullscreen={isFullscreen}
+              setIsFullscreen={setIsFullscreen}
+              onError={onError}
+              setIsCommentSectionOpen={setIsCommentSectionOpen}
+              sharedSlidesLink={sharedSlidesLink}
+              isPreviewMode={isPreviewMode}
+              documentName={documentName as string}
+            />
+          )}
           <div
-            id="toolbar"
             className={cn(
-              'z-50 hidden xl:flex items-center justify-center w-full h-[52px] fixed left-0 px-1 bg-[#ffffff] border-b color-border-default transition-transform duration-300 top-[3.5rem]',
+              'bg-white w-full mx-auto rounded',
+              { 'mt-4 md:!mt-16': isPreviewMode },
+              { 'md:!mt-16': !isPreviewMode },
+              { 'pt-20 md:!mt-[7.5rem]': isNavbarVisible && !isPreviewMode },
+              { 'pt-6 md:!mt-16': !isNavbarVisible && !isPreviewMode },
               {
-                'translate-y-0': isNavbarVisible,
-                'translate-y-[-108%]': !isNavbarVisible,
+                'max-[1080px]:!mx-auto min-[1081px]:!ml-[10%] min-[1700px]:!mx-auto':
+                  (isCommentSectionOpen || commentDrawerOpen) &&
+                  !isNativeMobile &&
+                  zoomLevel !== '0.5' &&
+                  zoomLevel !== '0.75' &&
+                  zoomLevel !== '1.4' &&
+                  zoomLevel !== '1.5' &&
+                  zoomLevel !== '2',
               },
+              {
+                '!mx-auto':
+                  !(isCommentSectionOpen || commentDrawerOpen) ||
+                  zoomLevel === '0.5' ||
+                  zoomLevel === '0.75' ||
+                  zoomLevel === '1.4' ||
+                  zoomLevel === '1.5',
+              },
+              { '!ml-0': zoomLevel === '2' && isWidth1500px && !isWidth3000px },
+              { 'min-h-[83vh]': isNavbarVisible },
+              { 'min-h-[90vh]': !isNavbarVisible },
+              { 'w-[700px] md:max-w-[700px] h-[150%]': zoomLevel === '0.5' },
+              { 'w-[800px] md:max-w-[800px] h-[200%]': zoomLevel === '0.75' },
+              { 'w-[850px] md:max-w-[850px] h-[100%]': zoomLevel === '1' },
+              { 'w-[70%] md:max-w-[70%] h-[200%]': zoomLevel === '1.4' },
+              {
+                'w-[1062.5px] md:max-w-[1062.5px] h-[100%]':
+                  zoomLevel === '1.5',
+              },
+              { 'w-[1548px] md:max-w-[1548px]': zoomLevel === '2' },
             )}
+            style={{
+              transformOrigin:
+                zoomLevel === '2' && !isWidth3000px
+                  ? 'left center'
+                  : 'top center',
+              transform: `scaleX(${zoomLevel})`,
+            }}
           >
-            <div className="justify-center items-center grow relative">
-              <EditorToolBar
+            <div
+              ref={editorRef}
+              className={cn(
+                'w-full h-full pt-8 md:pt-0',
+                { 'custom-ios-padding': isIOS },
+                { 'bg-white': zoomLevel === '1.4' || '1.5' },
+              )}
+              style={{
+                transformOrigin: 'top center',
+                transform: `scaleY(${zoomLevel})`,
+              }}
+            >
+              <div>
+                <EditorBubbleMenu
+                  editor={editor}
+                  onError={onError}
+                  zoomLevel={zoomLevel}
+                  setIsCommentSectionOpen={setIsCommentSectionOpen}
+                  inlineCommentData={inlineCommentData}
+                  setInlineCommentData={setInlineCommentData}
+                  isPreviewMode={isPreviewMode}
+                  username={username as string}
+                  walletAddress={walletAddress as string}
+                  onInlineComment={onInlineComment}
+                  setCommentDrawerOpen={setCommentDrawerOpen}
+                  activeCommentId={activeCommentId}
+                />
+                <ColumnsMenu editor={editor} appendTo={editorRef} />
+              </div>
+              <EditingProvider isPreviewMode={isPreviewMode}>
+                {tags && tags.length > 0 && (
+                  <div
+                    ref={tagsContainerRef}
+                    className={cn(
+                      'flex flex-wrap px-4 md:px-[80px] lg:!px-[124px] items-center gap-1 mb-4 mt-4 lg:!mt-0',
+                      { 'pt-12': isPreviewMode },
+                    )}
+                  >
+                    {visibleTags.map((tag, index) => (
+                      <Tag
+                        key={index}
+                        style={{ backgroundColor: tag?.color }}
+                        onRemove={() => handleRemoveTag(tag?.name)}
+                        isRemovable={!isPreviewMode}
+                        className="!h-6 rounded"
+                      >
+                        {tag?.name}
+                      </Tag>
+                    ))}
+                    {hiddenTagsCount > 0 && !isHiddenTagsVisible && (
+                      <Button
+                        variant="ghost"
+                        className="!h-6 rounded min-w-fit !px-2 color-bg-secondary text-helper-text-sm"
+                        onClick={() => setIsHiddenTagsVisible(true)}
+                      >
+                        +{hiddenTagsCount}
+                      </Button>
+                    )}
+                    <AnimatePresence>
+                      {isHiddenTagsVisible && (
+                        <motion.div
+                          initial={{ opacity: 0, height: 0 }}
+                          animate={{ opacity: 1, height: 'auto' }}
+                          exit={{ opacity: 0, height: 0 }}
+                          transition={{ duration: 0.3 }}
+                          className="flex flex-wrap items-center gap-1"
+                        >
+                          {selectedTags?.slice(4).map((tag, index) => (
+                            <Tag
+                              key={index + 4}
+                              style={{ backgroundColor: tag?.color }}
+                              onRemove={() => handleRemoveTag(tag?.name)}
+                              isRemovable={!isPreviewMode}
+                              className="!h-6 rounded"
+                            >
+                              {tag?.name}
+                            </Tag>
+                          ))}
+                        </motion.div>
+                      )}
+                    </AnimatePresence>
+                    {selectedTags && selectedTags?.length < 6 ? (
+                      <TagInput
+                        tags={tags || []}
+                        selectedTags={selectedTags as TagType[]}
+                        onAddTag={handleAddTag}
+                        isPreviewMode={isPreviewMode}
+                      />
+                    ) : null}
+                  </div>
+                )}
+                <EditorContent
+                  editor={editor}
+                  id="editor"
+                  className="w-full h-auto py-4 bg-white"
+                />
+              </EditingProvider>
+            </div>
+            {showCommentButton && !isNativeMobile && (
+              <Button
+                ref={btn_ref}
+                onClick={() => {
+                  handleCommentButtonClick?.(editor);
+                }}
+                variant="ghost"
+                className={cn(
+                  'absolute w-12 h-12 bg-white rounded-full shadow-xl top-[70px] right-[-23px]',
+                )}
+              >
+                <LucideIcon name="MessageSquareText" size="sm" />
+              </Button>
+            )}
+          </div>
+          {!isPreviewMode && !disableBottomToolbar && (
+            <div
+              className={cn(
+                'flex xl:hidden items-center w-full h-[52px] absolute left-0 z-40 px-4 bg-[#ffffff] transition-all duration-300 ease-in-out border-b border-color-default',
+                isKeyboardVisible && 'hidden',
+                { 'top-14': isNavbarVisible, 'top-0': !isNavbarVisible },
+              )}
+            >
+              <MobileToolbar
                 onError={onError}
                 editor={editor}
-                zoomLevel={zoomLevel}
-                setZoomLevel={setZoomLevel}
+                isKeyboardVisible={isKeyboardVisible}
                 isNavbarVisible={isNavbarVisible}
                 setIsNavbarVisible={setIsNavbarVisible}
                 secureImageUploadUrl={secureImageUploadUrl}
-                onMarkdownExport={onMarkdownExport}
-                onMarkdownImport={onMarkdownImport}
               />
             </div>
-          </div>
-        )}
-        {isPresentationMode && (
-          <PresentationMode
-            editor={editor}
-            onClose={handleClosePresentationMode}
-            isFullscreen={isFullscreen}
-            setIsFullscreen={setIsFullscreen}
-            onError={onError}
-            setIsCommentSectionOpen={setIsCommentSectionOpen}
-            sharedSlidesLink={sharedSlidesLink}
-            isPreviewMode={isPreviewMode}
-            documentName={documentName as string}
+          )}
+          <CommentDrawer
+            isOpen={commentDrawerOpen as boolean}
+            onClose={() => setCommentDrawerOpen?.(false)}
+            isNavbarVisible={isNavbarVisible}
+            isPresentationMode={isPresentationMode as boolean}
+            activeCommentId={activeCommentId}
           />
-        )}
-        <div
-          className={cn(
-            'bg-white w-full mx-auto rounded',
-            { 'mt-4 md:!mt-16': isPreviewMode },
-            { 'md:!mt-16': !isPreviewMode },
-            { 'pt-20 md:!mt-[7.5rem]': isNavbarVisible && !isPreviewMode },
-            { 'pt-6 md:!mt-16': !isNavbarVisible && !isPreviewMode },
-            {
-              'max-[1080px]:!mx-auto min-[1081px]:!ml-[10%] min-[1700px]:!mx-auto':
-                (isCommentSectionOpen || commentDrawerOpen) &&
-                !isNativeMobile &&
-                zoomLevel !== '0.5' &&
-                zoomLevel !== '0.75' &&
-                zoomLevel !== '1.4' &&
-                zoomLevel !== '1.5' &&
-                zoomLevel !== '2',
-            },
-            {
-              '!mx-auto':
-                !(isCommentSectionOpen || commentDrawerOpen) ||
-                zoomLevel === '0.5' ||
-                zoomLevel === '0.75' ||
-                zoomLevel === '1.4' ||
-                zoomLevel === '1.5',
-            },
-            { '!ml-0': zoomLevel === '2' && isWidth1500px && !isWidth3000px },
-            { 'min-h-[83vh]': isNavbarVisible },
-            { 'min-h-[90vh]': !isNavbarVisible },
-            { 'w-[700px] md:max-w-[700px] h-[150%]': zoomLevel === '0.5' },
-            { 'w-[800px] md:max-w-[800px] h-[200%]': zoomLevel === '0.75' },
-            { 'w-[850px] md:max-w-[850px] h-[100%]': zoomLevel === '1' },
-            { 'w-[70%] md:max-w-[70%] h-[200%]': zoomLevel === '1.4' },
-            {
-              'w-[1062.5px] md:max-w-[1062.5px] h-[100%]': zoomLevel === '1.5',
-            },
-            { 'w-[1548px] md:max-w-[1548px]': zoomLevel === '2' },
-          )}
-          style={{
-            transformOrigin:
-              zoomLevel === '2' && !isWidth3000px
-                ? 'left center'
-                : 'top center',
-            transform: `scaleX(${zoomLevel})`,
-          }}
-        >
-          <div
-            ref={editorRef}
-            className={cn(
-              'w-full h-full pt-8 md:pt-0',
-              { 'custom-ios-padding': isIOS },
-              { 'bg-white': zoomLevel === '1.4' || '1.5' },
-            )}
-            style={{
-              transformOrigin: 'top center',
-              transform: `scaleY(${zoomLevel})`,
-            }}
-          >
+          {!isNativeMobile && (
             <div>
-              <EditorBubbleMenu
-                editor={editor}
-                onError={onError}
-                zoomLevel={zoomLevel}
-                setIsCommentSectionOpen={setIsCommentSectionOpen}
-                inlineCommentData={inlineCommentData}
-                setInlineCommentData={setInlineCommentData}
-                isPreviewMode={isPreviewMode}
-                username={username as string}
-                walletAddress={walletAddress as string}
-                onInlineComment={onInlineComment}
-                setComment={setComment}
-                comments={comments}
-                setComments={setComments}
-                activeCommentId={activeCommentId as string}
-                commentDrawerOpen={commentDrawerOpen}
-                setCommentDrawerOpen={setCommentDrawerOpen}
-              />
-              <ColumnsMenu editor={editor} appendTo={editorRef} />
+              <CommentBubbleMenu editor={editor} zoomLevel={zoomLevel} />
             </div>
-            <EditingProvider isPreviewMode={isPreviewMode}>
-              {tags && tags.length > 0 && (
-                <div
-                  ref={tagsContainerRef}
-                  className={cn(
-                    'flex flex-wrap px-4 md:px-[80px] lg:!px-[124px] items-center gap-1 mb-4 mt-4 lg:!mt-0',
-                    { 'pt-12': isPreviewMode },
-                  )}
-                >
-                  {visibleTags.map((tag, index) => (
-                    <Tag
-                      key={index}
-                      style={{ backgroundColor: tag?.color }}
-                      onRemove={() => handleRemoveTag(tag?.name)}
-                      isRemovable={!isPreviewMode}
-                      className="!h-6 rounded"
-                    >
-                      {tag?.name}
-                    </Tag>
-                  ))}
-                  {hiddenTagsCount > 0 && !isHiddenTagsVisible && (
-                    <Button
-                      variant="ghost"
-                      className="!h-6 rounded min-w-fit !px-2 color-bg-secondary text-helper-text-sm"
-                      onClick={() => setIsHiddenTagsVisible(true)}
-                    >
-                      +{hiddenTagsCount}
-                    </Button>
-                  )}
-                  <AnimatePresence>
-                    {isHiddenTagsVisible && (
-                      <motion.div
-                        initial={{ opacity: 0, height: 0 }}
-                        animate={{ opacity: 1, height: 'auto' }}
-                        exit={{ opacity: 0, height: 0 }}
-                        transition={{ duration: 0.3 }}
-                        className="flex flex-wrap items-center gap-1"
-                      >
-                        {selectedTags?.slice(4).map((tag, index) => (
-                          <Tag
-                            key={index + 4}
-                            style={{ backgroundColor: tag?.color }}
-                            onRemove={() => handleRemoveTag(tag?.name)}
-                            isRemovable={!isPreviewMode}
-                            className="!h-6 rounded"
-                          >
-                            {tag?.name}
-                          </Tag>
-                        ))}
-                      </motion.div>
-                    )}
-                  </AnimatePresence>
-                  {selectedTags && selectedTags?.length < 6 ? (
-                    <TagInput
-                      tags={tags || []}
-                      selectedTags={selectedTags as TagType[]}
-                      onAddTag={handleAddTag}
-                      isPreviewMode={isPreviewMode}
-                    />
-                  ) : null}
-                </div>
-              )}
-              <EditorContent
-                editor={editor}
-                id="editor"
-                className="w-full h-auto py-4 bg-white"
-              />
-            </EditingProvider>
-          </div>
-          {showCommentButton && !isNativeMobile && (
-            <Button
-              ref={btn_ref}
-              onClick={() => {
-                handleCommentButtonClick?.(editor);
-              }}
-              variant="ghost"
-              className={cn(
-                'absolute w-12 h-12 bg-white rounded-full shadow-xl top-[70px] right-[-23px]',
-              )}
-            >
-              <LucideIcon name="MessageSquareText" size="sm" />
-            </Button>
           )}
-        </div>
-        {!isPreviewMode && !disableBottomToolbar && (
-          <div
-            className={cn(
-              'flex xl:hidden items-center w-full h-[52px] absolute left-0 z-50 px-4 bg-[#ffffff] transition-all duration-300 ease-in-out border-b border-color-default',
-              isKeyboardVisible && 'hidden',
-              { 'top-14': isNavbarVisible, 'top-0': !isNavbarVisible },
-            )}
-          >
-            <MobileToolbar
-              onError={onError}
-              editor={editor}
-              isKeyboardVisible={isKeyboardVisible}
-              isNavbarVisible={isNavbarVisible}
-              setIsNavbarVisible={setIsNavbarVisible}
-              secureImageUploadUrl={secureImageUploadUrl}
-            />
-          </div>
-        )}
-        <CommentDrawer
-          commentsSectionRef={commentsSectionRef}
-          isOpen={commentDrawerOpen as boolean}
-          onClose={() => setCommentDrawerOpen?.(false)}
-          comments={comments}
-          activeCommentId={activeCommentId}
-          username={username as string}
-          walletAddress={walletAddress as string}
-          editor={editor}
-          setComments={setComments}
-          setActiveCommentId={setActiveCommentId}
-          focusCommentInEditor={focusCommentInEditor}
-          handleAddReply={handleAddReply}
-          isNavbarVisible={isNavbarVisible}
-          isPresentationMode={isPresentationMode as boolean}
-        />
-        {!isNativeMobile && (
-          <div>
-            <CommentBubbleMenu
-              editor={editor}
-              comments={comments}
-              activeCommentId={activeCommentId}
-              zoomLevel={zoomLevel}
-              onPrevComment={() => {
-                const currentIndex = comments.findIndex(
-                  (comment) => comment.id === activeCommentId,
-                );
-                if (currentIndex > 0) {
-                  const prevComment = comments[currentIndex - 1];
-                  focusCommentInEditor(prevComment.id);
-                }
-              }}
-              onNextComment={() => {
-                const currentIndex = comments.findIndex(
-                  (comment) => comment.id === activeCommentId,
-                );
-                if (currentIndex < comments.length - 1) {
-                  const nextComment = comments[currentIndex + 1];
-                  focusCommentInEditor(nextComment.id);
-                }
-              }}
-            />
-          </div>
-        )}
+        </CommentProvider>
       </div>
     );
   },
